Adding more filters to app + update Readme

+++ compile.sh
@@ -1,6 +1,10 @@
 #! /bin/bash
 
-rm filter
-rm kernels.o
+if [ -e ./filter ] || [ -e ./kernels.o ];then
+    echo "Pre-existent files, excluding"
+    rm filter
+    rm kernels.o
+fi
+
 nvcc -c kernels.cu
 nvcc -ccbin g++ -Xcompiler "-std=c++11" kernels.o main.cpp lodepng.cpp helpers.cpp -lcuda -lcudart -o filter
